* (update) upgrade libosip2 to version 3.0.1

+++ wifo/libosip2/include/osipparser2/osip_md5.h
@@ -48,7 +48,7 @@ extern "C"
   typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-  typedef unsigned long int UINT4;
+  typedef unsigned int UINT4;
 
 /* PROTO_LIST is defined depending on how PROTOTYPES is defined above.
 If using PROTOTYPES, then PROTO_LIST returns the list, otherwise it
@@ -67,9 +67,9 @@ If using PROTOTYPES, then PROTO_LIST returns the list, otherwise it
  */
   typedef struct
   {
-    UINT4 state[4];		/* state (ABCD) */
-    UINT4 count[2];		/* number of bits, modulo 2^64 (lsb first) */
-    unsigned char buffer[64];	/* input buffer */
+    UINT4 state[4];             /* state (ABCD) */
+    UINT4 count[2];             /* number of bits, modulo 2^64 (lsb first) */
+    unsigned char buffer[64];   /* input buffer */
   }
   MD5_CTX;
